+++ tests/test54.c
@@ -15,7 +15,7 @@ int main()
     utstring_printf(t, "there " );
 
     utstring_concat(s, t);
-    printf("length: %u\n", utstring_len(s));
+    printf("length: %u\n", (unsigned)utstring_len(s));
     printf("%s\n", utstring_body(s));
 
     utstring_free(s);
