+++ tests/test56.c
@@ -87,7 +87,7 @@ int main(int argc, char *argv[])
 
     utstring_new(s);
     utstring_bincpy(s, binary, sizeof(binary));
-    printf("length is %u\n", utstring_len(s));
+    printf("length is %u\n", (unsigned)utstring_len(s));
 
     utstring_clear(s);
     utstring_printf(s,"number %d", 10);
